+++ kotsadm/web/src/components/apps/AppVersionHistory.jsx
@@ -16,8 +16,9 @@ import Loader from "../shared/Loader";
 import MarkdownRenderer from "@src/components/shared/MarkdownRenderer";
 import DownstreamWatchVersionDiff from "@src/components/watches/DownstreamWatchVersionDiff";
 import AirgapUploadProgress from "@src/components/AirgapUploadProgress";
+import UpdateCheckerModal from "@src/components/modals/UpdateCheckerModal";
 import { getKotsDownstreamHistory, getKotsDownstreamOutput, getUpdateDownloadStatus } from "../../queries/AppsQueries";
-import { Utilities, isAwaitingResults, secondsAgo, getPreflightResultState, getGitProviderDiffUrl, getCommitHashFromUrl, getReadableCronDescriptor } from "../../utilities/utilities";
+import { Utilities, isAwaitingResults, secondsAgo, getPreflightResultState, getGitProviderDiffUrl, getCommitHashFromUrl } from "../../utilities/utilities";
 import { Repeater } from "../../utilities/repeater";
 import has from "lodash/has";
 import get from "lodash/get";
@@ -55,9 +56,7 @@ class AppVersionHistory extends Component {
     updateChecker: new Repeater(),
     uploadTotal: 0,
     uploadSent: 0,
-    showUpdateCheckerModal: false,
-    updateCheckerSpec: "",
-    submitUpdateCheckerSpecErr: ""
+    showUpdateCheckerModal: false
   }
 
   componentDidMount() {
@@ -71,16 +70,6 @@ class AppVersionHistory extends Component {
       if (firstSequence !== undefined && secondSequence !== undefined) { // undefined because a sequence can be zero!
         this.setState({ showDiffOverlay: true, firstSequence, secondSequence });
       }
-    }
-
-    if (this.props.app) {
-      this.setState({ updateCheckerSpec: this.props.app.updateCheckerSpec });
-    }
-  }
-
-  componentDidUpdate(lastProps) {
-    if (this.props.app && this.props.app?.updateCheckerSpec !== lastProps.app?.updateCheckerSpec) {
-      this.setState({ updateCheckerSpec: this.props.app.updateCheckerSpec });
     }
   }
 
@@ -122,61 +111,6 @@ class AppVersionHistory extends Component {
     this.setState({
       showUpdateCheckerModal: true
     });
-  }
-
-  onSubmitUpdateCheckerSpec = () => {
-    const { updateCheckerSpec } = this.state;
-    const { app } = this.props;
-
-    this.setState({
-      submitUpdateCheckerSpecErr: ""
-    });
-
-    fetch(`${window.env.API_ENDPOINT}/app/${app.slug}/updatecheckerspec`, {
-      headers: {
-        "Authorization": Utilities.getToken(),
-        "Content-Type": "application/json",
-      },
-      method: "PUT",
-      body: JSON.stringify({
-        updateCheckerSpec: updateCheckerSpec,
-      })
-    })
-      .then(async (res) => {
-        if (!res.ok) {
-          const text = await res.text();
-          this.setState({
-            submitUpdateCheckerSpecErr: text
-          });
-          return;
-        }
-
-        this.setState({
-          submitUpdateCheckerSpecErr: "",
-          showUpdateCheckerModal: false
-        });
-        
-        this.props.refreshAppData();
-      })
-      .catch((err) => {
-        this.setState({
-          submitUpdateCheckerSpecErr: String(err)
-        });
-      });
-  }
-
-  getReadableCronExpression = () => {
-    const { updateCheckerSpec } = this.state;
-    try {
-      const readable = getReadableCronDescriptor(updateCheckerSpec);
-      if (readable.includes("undefined")) {
-        return "";
-      } else {
-        return readable;
-      }
-    } catch(error) {
-      return "";
-    }
   }
 
   getVersionDiffSummary = version => {
@@ -821,8 +755,6 @@ class AppVersionHistory extends Component {
       uploadSent,
       noUpdateAvailiableText,
       showUpdateCheckerModal,
-      updateCheckerSpec,
-      submitUpdateCheckerSpecErr,
     } = this.state;
 
     if (!app) {
@@ -906,7 +838,6 @@ class AppVersionHistory extends Component {
       data?.stopPolling();
     }
 
-    const humanReadableCron = this.getReadableCronExpression(updateCheckerSpec);
 
     return (
       <div className="flex flex-column flex1 u-position--relative u-overflow--auto u-padding--20">
@@ -1220,38 +1151,18 @@ class AppVersionHistory extends Component {
           </div>
         </Modal>
 
-        <Modal
-          isOpen={showUpdateCheckerModal}
-          onRequestClose={this.hideUpdateCheckerModal}
-          contentLabel="Update Checker"
-          ariaHideApp={false}
-          className="Modal"
-        >
-          <div className="u-position--relative flex-column u-padding--20">
-            <span className="u-fontSize--largest u-fontWeight--bold u-color--tuna u-marginBottom--15">Configure update checker</span>
-            <div className="info-box u-marginBottom--20">
-              <span className="u-fontSize--small u-textAlign--center">
-                You can enter <span className="u-fontWeight--bold u-color--tuna">@never</span> to disable scheduled update checks
-              </span>
-            </div>
-            <div className="flex-column flex1 u-paddingLeft--5">
-              <p className="u-fontSize--normal u-color--tuna u-fontWeight--bold u-lineHeight--normal u-marginBottom--10">Cron expression</p>
-              <input
-                type="text"
-                className="Input u-marginBottom--5"
-                placeholder="0 0 * * MON"
-                value={updateCheckerSpec}
-                onChange={(e) => this.setState({ updateCheckerSpec: e.target.value })}
-              />
-              {humanReadableCron && <span className="u-fontSize--small u-fontWeight--medium u-color--dustyGray">{humanReadableCron}</span>}
-              {submitUpdateCheckerSpecErr && <span className="u-color--chestnut u-fontSize--small u-fontWeight--bold u-marginTop--15">{submitUpdateCheckerSpecErr}</span>}
-            </div>
-            <div className="flex u-marginTop--20">
-              <button className="btn primary blue" onClick={this.onSubmitUpdateCheckerSpec}>Update</button>
-              <button className="btn secondary u-marginLeft--10" onClick={this.hideUpdateCheckerModal}>Cancel</button>
-            </div>
-          </div>
-        </Modal>
+        {showUpdateCheckerModal &&
+          <UpdateCheckerModal
+            isOpen={showUpdateCheckerModal}
+            onRequestClose={this.hideUpdateCheckerModal}
+            updateCheckerSpec={app.updateCheckerSpec}
+            appSlug={app.slug}
+            onUpdateCheckerSpecSubmitted={() => {
+              this.hideUpdateCheckerModal();
+              this.props.refreshAppData();
+            }}
+          />
+        }
       </div>
     );
   }
